Updated Melee attack System new issue targets changing as character moves.
