+++ Assets/Systems/IAUS/Components/Systems/Reactive System/Attack SubState Reactive.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 15a20bcbb7b5311488c42ec8c3e9fa6f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
