+++ Assets/Systems/IAUS/Components/Systems/Reactive System/Attack SubState Reactive/MeleeReactive.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e244ca805ed8b414eb502fb6c120103f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
